Line (Timeline) Map
To work with the schedule, some functionality is needed that will allow
processing and storing a set of events distributed in time. Therefore,
something is needed that could be presented as a time line on which
events of different lengths can be placed.

The Line Map is a line on which some segments can be placed on this
line. It arranges segments in sorted form and provides quick access to
added segments by the requested range.

The Line Map implements an interface similar to the standard Map
interface, where any value (both objects and primitive values) can be
used as a key, and the value is a range or a point on a line.

The Line Map intermnally implements a AVL tree. AVL tree was chosen
because it provides a faster value lookup.