Schedule
The Schedule is a public feature of the library that provides the
ability to schedule, unschedule, and reschedule an event.

When scheduling an event, the user receives a inique digital ID for that
schedule, which the user can use to unschedule and reschedule the event,
as well as to link to their data.